move MainViewer/ItemTypes.ts -> constants/DndItemTypes.ts
