--- src/renderer/prototypes/MainViewer/ItemTypes.ts
@@ -1,1 +1,0 @@
-export const DndItemTypes = {TRACK: "track"};
