+++ src/renderer/prototypes/constants/DndItemTypes.ts
@@ -1,0 +1,1 @@
+export default {TRACK: "track"};
